backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
databases[postgresql]==0.8.0
sqlalchemy==1.4.51
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.0
